Relocate SvcLoad functions
