+++ components/hab/src/cli/hab/svc.rs
@@ -267,3 +267,76 @@ pub fn svc_loads_from_paths<T: AsRef<Path>>(paths: &[T]) -> Result<Vec<Load>> {
     }
     Ok(svc_loads)
 }
+
+pub fn shared_load_cli_to_ctl(ident: PackageIdent,
+                              shared_load: SharedLoad,
+                              force: bool)
+                              -> habitat_sup_protocol::ctl::SvcLoad {
+    use habitat_common::{ui,
+                         ui::UIWriter};
+    use habitat_sup_protocol::{ctl::{ServiceBindList,
+                                     SvcLoad},
+                               types::{HealthCheckInterval,
+                                       ServiceBind}};
+
+    // TODO (DM): This check can eventually be removed.
+    // See https://github.com/habitat-sh/habitat/issues/7339
+    if !shared_load.application.is_empty() || !shared_load.environment.is_empty() {
+        ui::ui().warn("--application and --environment flags are deprecated and ignored.")
+                .ok();
+    }
+
+    let binds = if shared_load.bind.is_empty() {
+        None
+    } else {
+        Some(ServiceBindList { binds: shared_load.bind
+                                                 .into_iter()
+                                                 .map(ServiceBind::from)
+                                                 .collect(), })
+    };
+
+    let config_from = if let Some(config_from) = shared_load.config_from {
+        warn!("");
+        warn!("WARNING: Setting '--config-from' should only be used in development, not \
+               production!");
+        warn!("");
+        Some(config_from.to_string_lossy().to_string())
+    } else {
+        None
+    };
+
+    #[cfg(target_os = "windows")]
+    let svc_encrypted_password = if let Some(password) = shared_load.password {
+        Some(encrypt(password)?)
+    } else {
+        None
+    };
+    #[cfg(not(target_os = "windows"))]
+    let svc_encrypted_password = None;
+
+    SvcLoad { ident: Some(ident.into()),
+              application_environment: None,
+              binds,
+              specified_binds: None,
+              binding_mode: Some(shared_load.binding_mode as i32),
+              bldr_url: Some(habitat_core::url::bldr_url(shared_load.bldr_url)),
+              bldr_channel: Some(shared_load.channel.to_string()),
+              config_from,
+              force: Some(force),
+              group: Some(shared_load.group),
+              svc_encrypted_password,
+              topology: shared_load.topology.map(i32::from),
+              update_strategy: Some(shared_load.strategy as i32),
+              health_check_interval:
+                  Some(HealthCheckInterval { seconds: shared_load.health_check_interval, }),
+              shutdown_timeout: shared_load.shutdown_timeout.map(u32::from),
+              update_condition: Some(shared_load.update_condition as i32) }
+}
+
+impl From<Load> for habitat_sup_protocol::ctl::SvcLoad {
+    fn from(svc_load: Load) -> Self {
+        shared_load_cli_to_ctl(svc_load.pkg_ident.pkg_ident(),
+                               svc_load.shared_load,
+                               svc_load.force)
+    }
+}
